% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sigmaplot.R
\name{theme_sigmaplot}
\alias{theme_sigmaplot}
\title{Sigmaplot styled theme for Ggplot2}
\usage{
theme_sigmaplot(base_size = 16)
}
\arguments{
\item{base_size}{}
}
\value{
a sigmaplot styled ggplot2 object
}
\description{
Sigmaplot styled theme for Ggplot2
}
\examples{
plot <- ggplot(mpg, aes(displ, hwy)) + geom_point()
plot + theme_sigmaplot()
}
